HTCONDOR-3540: Restore SimScheddQ and manually 'simulate' ad creation in checkJDL
